+++ WasteIntoCity/WasteIntoCity.Api/Contracts/V1/Responses/WorkGetByIdResponse.cs
@@ -16,8 +16,8 @@
 
         required public int WorkStatusTypesId { get; init; }
 
-        required public double Lat { get; init; }
+        required public decimal Lat { get; init; }
 
-        required public double Lng { get; init; }
+        required public decimal Lng { get; init; }
     }
 }
