+++ WasteIntoCity/WasteIntoCity.Core/Interfaces/Services/IWorkApplicationsService.cs
@@ -2,7 +2,7 @@
 {
     public interface IWorkApplicationsService
     {
-        Task CreateOwnAsync(string title, string description, int workComplexityId, double lat, double lng, Guid userId);
+        Task CreateOwnAsync(string title, string description, int workComplexityId, decimal lat, decimal lng, Guid userId);
 
         Task RejectAsync(Guid workApplicationsId);
